requirements.txt
fastapi>=0.114.0
starlette>=0.38.5
uvicorn[standard]
pytest
requests
httpx>=0.27
python-dotenv
cryptography
pymysql
